+++ priv/repo/migrations/20221115193254_create_products.exs
@@ -5,7 +5,7 @@ defmodule Horizon65.Repo.Migrations.CreateProducts do
     create table(:products) do
       add :code, :string
       add :name, :string
-      add :price, :integer
+      add :price, :float
 
       timestamps()
     end
